apex/apex-agents/requirements.txt
# Requirements for apex-agents
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.11.7
asyncpg==0.30.0
python-dotenv==1.0.0
google-genai==1.66.0
pytest==7.4.0
pytest-asyncio==0.21.0
